+++ lib/cdk/applyHttpInterceptor.ts
@@ -3,7 +3,7 @@ import { Aspects, IAspect } from 'aws-cdk-lib';
 import { applyNodejsInternalExtensionToNodeJsFunction } from './applyNodejsInternalExtension';
 import { HttpInterceptor } from './HttpInterceptor';
 import { NodejsFunction } from 'aws-cdk-lib/aws-lambda-nodejs';
-import { HTTP_INTERCEPTOR_TABLE_NAME } from './sdk';
+import { HTTP_INTERCEPTOR_TABLE_NAME } from '../sdk';
 
 class HttpInterceptorApplier implements IAspect {
   constructor(private httpInterceptor: HttpInterceptor) {}
