+++ lib/cdk/applyNodejsInternalExtension.ts
